Fixed a panic that would occur on certain times
I found a bug where certain times provided by the `time` crate would cause a crash.  The first thing i did was bump the version number on the `time` crate.  I immediately realized that the crate had changed drastically; I believe it can no longer be used in this application, so I switched to the `chrono` crate.  The `chrono` crate is more user friendly and offered methods that allowed me to reduce the time-conversion logic that was the source of the original crash, so I believe the bug is fixed now just from swapping out the `time` crate to the `chrono` crate.